Moved location tracking logic to view controller
